# docs/public/openapi.yaml
openapi: 3.0.3
info:
  title: TeslaSync API
  version: 1.0.0
  description: Self-hosted Tesla Fleet Intelligence API
  license:
    name: MIT
  contact:
    name: TeslaSync
    url: https://github.com/ev-dev-labs/teslasync
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /export/jobs/import:
    post:
      operationId: submitImportJob
      summary: Submit import job
      description: Submits a data import job.
      tags: [Export]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Import job submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    example: pending
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            invalidInput:
              value:
                error: invalid input parameters
                code: VALIDATION_INVALID_INPUT
                category: validation
            invalidJson:
              value:
                error: invalid JSON body
                code: VALIDATION_INVALID_JSON
                category: validation
            missingField:
              value:
                error: required field missing
                code: VALIDATION_MISSING_FIELD
                category: validation
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: an internal error occurred
            code: INTERNAL_ERROR
            category: internal
  schemas:
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: resource not found
        code:
          type: string
          description: Machine-readable error code
          example: DB_RECORD_NOT_FOUND
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - UNPROCESSABLE_ENTITY
            - AUTH_INVALID_CREDENTIALS
            - AUTH_TOKEN_EXPIRED
            - AUTH_TOKEN_INVALID
            - AUTH_TOKEN_MISSING
            - AUTH_API_KEY_INVALID
            - AUTH_API_KEY_EXPIRED
            - AUTH_API_KEY_REVOKED
            - AUTH_INSUFFICIENT_PERMISSIONS
            - VEHICLE_NOT_FOUND
            - VEHICLE_OFFLINE
            - VEHICLE_ASLEEP
            - VEHICLE_BUSY
            - DRIVE_NOT_FOUND
            - CHARGE_NOT_FOUND
            - TRIP_NOT_FOUND
            - SESSION_STILL_ACTIVE
            - SESSION_NOT_FOUND
            - TESLA_API_UNAVAILABLE
            - TESLA_API_SUSPENDED
            - TESLA_API_RATE_LIMITED
            - TESLA_API_TIMEOUT
            - TESLA_API_AUTH_FAILED
            - TESLA_NOT_CONNECTED
            - DB_CONNECTION_FAILED
            - DB_QUERY_FAILED
            - DB_TRANSACTION_FAILED
            - DB_RECORD_NOT_FOUND
            - DB_DUPLICATE_RECORD
            - VALIDATION_INVALID_INPUT
            - VALIDATION_INVALID_JSON
            - VALIDATION_MISSING_FIELD
            - VALIDATION_INVALID_RANGE
            - VALIDATION_INVALID_ID
            - VALIDATION_PAYLOAD_TOO_LARGE
            - GEOFENCE_NOT_FOUND
            - GEOFENCE_INVALID_COORDINATES
            - GEOFENCE_INVALID_RADIUS
            - COMMAND_NOT_SUPPORTED
            - COMMAND_FAILED
            - COMMAND_TIMEOUT
            - BACKUP_CONFIG_NOT_FOUND
            - BACKUP_RUN_NOT_FOUND
            - BACKUP_FAILED
            - BACKUP_STORAGE_ERROR
            - RESTORE_FAILED
            - NOTIFICATION_CHANNEL_NOT_FOUND
            - NOTIFICATION_TEST_FAILED
            - TELEMETRY_INGEST_FAILED
            - MQTT_UNAVAILABLE
            - EXPORT_FAILED
            - EXPORT_NOT_FOUND
            - EXPORT_INVALID_FORMAT
            - RATE_LIMITED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - GATEWAY_TIMEOUT
            - NOT_IMPLEMENTED
            - ERROR
        category:
          type: string
          description: Error category
          example: database
          enum:
            - auth
            - vehicle
            - drives
            - charging
            - trips
            - tesla_api
            - database
            - validation
            - geofence
            - commands
            - backup
            - notifications
            - telemetry
            - export
            - rate_limiting
            - internal
...